% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_elements.R
\name{find_elements}
\alias{find_elements}
\title{Find the path from root to highest occurrence of each element}
\usage{
find_elements(hc)
}
\arguments{
\item{hc}{a protoclust object}
}
\value{
\item{paths}{a list of length n giving, for each element, the path from
root to its highest occurrence. A 0 means go left, a 1 means go right.}
\item{int_paths}{a list of length n - 1 giving, for each interior node,
the path from root to it. A 0 means go left, a 1 means go right.}
}
\description{
A protoclust object has a prototype associated with each interior node.
Every element being clustered occurs at least as a leaf but might also 
appear multiple times as a prototype.  This function finds for each element
the path from the root to the highest occurrence of that element.  The path
is specified by a series of 0s and 1s, where 0 means "go left" and 1 means
"go right".
}
