% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{plot.protoclust}
\alias{plot.protoclust}
\title{Plot Dendrogram}
\usage{
\method{plot}{protoclust}(x, ...)
}
\description{
Calls \code{\link{plotwithprototypes}}, which allows one to add prototype
labels to the dendrogram.
}

