% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfunctions.R
\name{plot.protoclust}
\alias{plot.protoclust}
\title{Plot Dendrogram}
\usage{
\method{plot}{protoclust}(x, ...)
}
\arguments{
\item{x}{a protoclust object}

\item{...}{additional arguments to be passed to 
\code{\link{plotwithprototypes}}}
}
\description{
Calls \code{\link{plotwithprototypes}}, which allows one to add prototype
labels to the dendrogram.
}
